+++ crates/ruma-federation-api/src/membership/create_join_event/v1.rs
@@ -65,13 +65,9 @@ impl Response {
 }
 
 /// Full state of the room.
-#[derive(Clone, Debug, Deserialize, Serialize)]
+#[derive(Clone, Debug, Default, Deserialize, Serialize)]
 #[cfg_attr(not(ruma_unstable_exhaustive_types), non_exhaustive)]
 pub struct RoomState {
-    #[cfg(not(feature = "unstable-unspecified"))]
-    /// The resident server's DNS name.
-    pub origin: String,
-
     /// The full set of authorization events that make up the state of the room,
     /// and their authorization events, recursively.
     pub auth_chain: Vec<Box<RawJsonValue>>,
@@ -87,29 +83,9 @@ pub struct RoomState {
     pub event: Option<Box<RawJsonValue>>,
 }
 
-#[cfg(feature = "unstable-unspecified")]
-impl Default for RoomState {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
 impl RoomState {
-    #[cfg(not(feature = "unstable-unspecified"))]
-    /// Creates an empty `RoomState` with the given `origin`.
-    ///
-    /// With the `unstable-unspecified` feature, this method doesn't take any parameters.
-    /// See [matrix-spec#374](https://github.com/matrix-org/matrix-spec/issues/374).
-    pub fn new(origin: String) -> Self {
-        Self { origin, auth_chain: Vec::new(), state: Vec::new(), event: None }
-    }
-
-    #[cfg(feature = "unstable-unspecified")]
-    /// Creates an empty `RoomState` with the given `origin`.
-    ///
-    /// Without the `unstable-unspecified` feature, this method takes a parameter for the origin.
-    /// See [matrix-spec#374](https://github.com/matrix-org/matrix-spec/issues/374).
+    /// Creates an empty `RoomState`.
     pub fn new() -> Self {
-        Self { auth_chain: Vec::new(), state: Vec::new(), event: None }
+        Self::default()
     }
 }